made enum transport as string
