+++ BrokerageApi/V1/Infrastructure/AuditEvents/AuditEventType.cs
@@ -1,0 +1,18 @@
+using Newtonsoft.Json;
+using Newtonsoft.Json.Converters;
+
+namespace BrokerageApi.V1.Infrastructure.AuditEvents
+{
+    [JsonConverter(typeof(StringEnumConverter))]
+    public enum AuditEventType
+    {
+        ReferralBrokerAssignment,
+        ReferralBrokerReassignment,
+        ElementEnded,
+        ElementCancelled,
+        ElementSuspended,
+        CarePackageEnded,
+        CarePackageCancelled,
+        CarePackageSuspended
+    }
+}
